Refactor: use size_t for FLAG_LENGTH

+++ argument_parser.cpp
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <string.h>
 
-static void getFieldWidth(const Flag flags[], int nFlags, size_t *fullNameFieldWidth, size_t *aliasFieldWidth);
+static void getFieldWidth(const Flag flags[], size_t nFlags, size_t *fullNameFieldWidth, size_t *aliasFieldWidth);
 
 static bool IsEqualFlag(char *arg, const char *flagName) {
     assert(arg);
@@ -15,17 +15,17 @@ static bool IsEqualFlag(char *arg, const char *flagName) {
     return strcmp(arg, flagName) == 0;
 }
 
-static const Flag *GetFlag(char *arg, const Flag flags[], int nFlags) {
+static const Flag *GetFlag(char *arg, const Flag flags[], size_t nFlags) {
     assert(arg);
     assert(flags);
     
-    for (int i = 0; i < nFlags; i++)
+    for (size_t i = 0; i < nFlags; i++)
         if (IsEqualFlag(arg, flags[i].fullName) || IsEqualFlag(arg, flags[i].alias))
             return flags+i;
     return NULL;
 }
 
-ParseCode ParseFlags(char *argv[], int argc, const Flag flags[], int nFlags, void *context) {
+ParseCode ParseFlags(char *argv[], int argc, const Flag flags[], size_t nFlags, void *context) {
     assert(argv != NULL);
     assert(flags != NULL);
 
@@ -50,23 +50,23 @@ ParseCode ParseFlags(char *argv[], int argc, const Flag flags[], int nFlags, voi
     return PC_NO_ERROR_CONTINUE;
 }
 
-void PrintArgumentInfo(const Flag flags[], int nFlags) {
+void PrintArgumentInfo(const Flag flags[], size_t nFlags) {
     assert(flags != NULL);
 
     size_t fullNameFieldWidth = 0;
     size_t aliasFieldWidth = 0;
     getFieldWidth(flags, nFlags, &fullNameFieldWidth, &aliasFieldWidth);
 
-    for (int i = 0; i < nFlags; i++)
+    for (size_t i = 0; i < nFlags; i++)
         printf(" --%-*s -%-*s %s\n", (int)fullNameFieldWidth, flags[i].fullName, (int)aliasFieldWidth, flags[i].alias, flags[i].description);
 }
 
-static void getFieldWidth(const Flag flags[], int nFlags, size_t *fullNameFieldWidth, size_t *aliasFieldWidth) {
+static void getFieldWidth(const Flag flags[], size_t nFlags, size_t *fullNameFieldWidth, size_t *aliasFieldWidth) {
     assert(flags);
     assert(fullNameFieldWidth);
     assert(aliasFieldWidth);
 
-    for (int i = 0; i < nFlags; i++) {
+    for (size_t i = 0; i < nFlags; i++) {
         if (*fullNameFieldWidth < strlen(flags[i].fullName))
             *fullNameFieldWidth = strlen(flags[i].fullName);
         if (*aliasFieldWidth < strlen(flags[i].alias))
